% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONI.R
\name{buildRegression}
\alias{buildRegression}
\title{Create one regression functions with one vector of predictor genes}
\usage{
buildRegression(ControlRNA, gene, gene_list)
}
\arguments{
\item{ControlRNA}{wild type control RNA seq counts where columns are samples and rows are genes}

\item{gene}{A character string naming the gene of interest}

\item{gene_list}{A character vector of genes to use as predictors}
}
\value{
A linear regression model
}
\description{
Create one regression functions with one vector of predictor genes
}
