% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATonI.R
\name{evaluateMutants}
\alias{evaluateMutants}
\title{Evaluate variant harbored by mutant samples}
\usage{
evaluateMutants(Function, ImmuneProb, ControlRNA)
}
\arguments{
\item{Function}{A data.frame returned by getMutantFunction}

\item{ImmuneProb}{A data.frame returned by getImmuneProb}

\item{ControlRNA}{wild type control RNA seq counts where columns are samples and rows are genes}
}
\value{
A list with the elements Function (function annotation of variant), p.val.func (p value of function annotation), ImmunePhenotype (immune phenotype annotation of variant), and p.val.immune (p value of immune phenotype annotation)
}
\description{
Evaluate variant harbored by mutant samples
}
