% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_huethr_values.R
\name{get_huethr_values}
\alias{get_huethr_values}
\title{Plot the histogram of hue values of image to determine the hue range of the background}
\usage{
get_huethr_values(img, return.values = TRUE)
}
\arguments{
\item{return.values}{}
}
\description{
Plot the histogram of hue values of image to determine the hue range of the background
}
