% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATonI.R
\name{getMutantFunction}
\alias{getMutantFunction}
\title{Predict function of gene of interest within mutant samples}
\usage{
getMutantFunction(Regression, MasterRNA, gene)
}
\arguments{
\item{Regression}{mutantRegression output}

\item{MasterRNA}{mutated samples' RNA seq counts where columns are samples and rows are genes}

\item{gene}{A character string naming the gene of interest}
}
\value{
A data.frame with sample names (taken from MasterRNA column names) and the predicted function of gene of interest
}
\description{
Predict function of gene of interest within mutant samples
}
