% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATonI.R
\name{getRegression}
\alias{getRegression}
\title{Create two regression functions with the vectors from makeGeneList}
\usage{
getRegression(gene_list_ppi, gene, ControlRNA)
}
\arguments{
\item{gene_list_ppi}{makeGeneList output}

\item{gene}{A character string naming the gene of interest}

\item{ControlRNA}{wild type control RNA seq counts where columns are samples and rows are genes}
}
\value{
A list of two linear regression models
}
\description{
Create two regression functions with the vectors from makeGeneList
}
