% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernl_to_coo_list}
\alias{kernl_to_coo_list}
\title{Convert a krnel object into a coo list from Momocs2 package}
\usage{
kernl_to_coo_list(x)
}
\arguments{
\item{x}{a krnel object as returned by krnel function}
}
\value{
a coo list
}
\description{
Convert a krnel object into a coo list from Momocs2 package
}
