% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONI.R
\name{makeGeneList}
\alias{makeGeneList}
\title{Create an upstream gene list, and a downstream gene list using STRING}
\usage{
makeGeneList(gene, sdb = sdb)
}
\arguments{
\item{gene}{A character string naming the gene of interest}

\item{sdb}{The list of protein interactions (provided)}
}
\value{
A list of two vectors with the upstream and downstream results
}
\description{
Create an upstream gene list, and a downstream gene list using STRING
}
