% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONI.R
\name{mutantRegression}
\alias{mutantRegression}
\title{Get prediction intervals of gene of interest counts for mutant samples}
\usage{
mutantRegression(Regression, MasterRNA, gene)
}
\arguments{
\item{Regression}{getRegression output}

\item{MasterRNA}{mutated samples' RNA seq counts where columns are samples and rows are genes}

\item{gene}{A character string naming the gene of interest}
}
\value{
A dataframe with the true counts and upper and lower 95\% prediction interval bounds
}
\description{
Get prediction intervals of gene of interest counts for mutant samples
}
