% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.krnel.R
\name{plot.krnel}
\alias{plot.krnel}
\title{plot features of a krnel object}
\usage{
\method{plot}{krnel}(x, label = TRUE, label.cex = 0.6,
  label.col = "red", plot.bbox = T, bbox.col = "blue", title = T,
  plot.radii = T, plot.ellipse = T, plot.ws.bbox = F,
  ws.bbox.col = "pink", plot.pois = F, plot.ws.pois = F,
  poi.col = "lightblue", poi.ws.col = "lightblue", custom.col = NULL)
}
\arguments{
\item{x}{a krnel object as returned by the krnel function}

\item{label}{boolean, plot label (ie number)}

\item{label.cex}{label size}

\item{label.col}{label color}

\item{plot.bbox}{whether to plot bounding boxes or not}

\item{bbox.col}{color of bounding box}
}
\value{
nothing
}
\description{
plot features of a krnel object
}
