% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONI.R
\name{getImmuneProb}
\alias{getImmuneProb}
\title{Predict immune phenotype within mutant samples}
\usage{
getImmuneProb(MasterRNA, lasso = lasso_best, model = m, geneLength = geneLen)
}
\arguments{
\item{MasterRNA}{mutated samples' RNA seq counts where columns are samples and rows are genes}

\item{lasso}{ElasticNet output to select genes correllated with immune phenotype (provided)}

\item{model}{OrdinalForest model to predict immune phenotype (provided)}

\item{geneLength}{Dataframe of gene lengths (provided)}
}
\value{
A data.frame with sample names (taken from MasterRNA column names) and the predicted immune phenotype
}
\description{
Predict immune phenotype within mutant samples
}
