% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATONI.R
\name{runRIGATONI}
\alias{runRIGATONI}
\title{Predict immune phenotype within mutant samples}
\usage{
runRIGATONI(
  gene,
  ControlRNA,
  MasterRNA,
  lasso_best = lasso_best,
  m = m,
  geneLen = geneLen,
  sdb = sdb
)
}
\arguments{
\item{gene}{A character string naming the gene of interest}

\item{ControlRNA}{wild type control RNA seq counts where columns are samples and rows are genes}

\item{MasterRNA}{mutated samples' RNA seq counts where columns are samples and rows are genes}

\item{lasso_best}{ElasticNet output to select genes correllated with immune phenotype (provided)}

\item{m}{OrdinalForest model to predict immune phenotype (provided)}

\item{geneLen}{Dataframe of gene lengths (provided)}

\item{sdb}{The list of protein interactions (provided)}
}
\value{
A data.frame with sample names (taken from MasterRNA column names), function prediction, and the predicted immune phenotype
}
\description{
Predict immune phenotype within mutant samples
}
