% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGATonI.R
\name{runRIGATonI}
\alias{runRIGATonI}
\title{Predict immune phenotype within mutant samples}
\usage{
runRIGATonI(
  gene,
  ControlRNA,
  MasterRNA,
  model_l = mod_l,
  model_h = mod_h,
  use_genes = genes,
  string = sdb,
  geneList = NULL
)
}
\arguments{
\item{gene}{A character string naming the gene of interest}

\item{ControlRNA}{wild type control RNA seq counts where columns are samples and rows are genes}

\item{MasterRNA}{mutated samples' RNA seq counts where columns are samples and rows are genes}

\item{model_l}{Model to predict Low immune advantage immune phenotype (provided)}

\item{model_h}{Model to predict High immune advantage immune phenotype (provided)}

\item{use_genes}{genes to predict Low immune advantage immune phenotype (provided)}

\item{string}{The list of protein interactions (default provided)}

\item{geneList}{list of upstream and downstream genes to use to predict mutant function. Default value is NULL and geneList will be built using STRING database.}
}
\value{
A data.frame with sample names (taken from MasterRNA column names), function prediction, and the predicted immune phenotype
}
\description{
Predict immune phenotype within mutant samples
}
